+++ tests/unittests/test_http_functions.py
@@ -8,7 +8,6 @@ import sys
 import typing
 from unittest import skipIf
 
-from azure_functions_worker.constants import SCRIPT_FILE_NAME
 from tests.utils import testutils
 
 
@@ -457,25 +456,3 @@ class TestHttpFunctionsStein(TestHttpFunctions):
         r = self.webhost.request('GET', 'no_return_returns')
         self.assertEqual(r.status_code, 200)
 
-
-class TestHttpFunctionsFileName(TestHttpFunctions):
-
-    @classmethod
-    def get_script_dir(cls):
-        return testutils.UNIT_TESTS_FOLDER / 'http_functions' / \
-                                             'http_functions_stein' / \
-                                             'file_name'
-
-    def test_no_return(self):
-        r = self.webhost.request('GET', 'no_return')
-        self.assertEqual(r.status_code, 500)
-
-    def test_no_return_returns(self):
-        r = self.webhost.request('GET', 'no_return_returns')
-        self.assertEqual(r.status_code, 200)
-
-    def test_correct_file_name(self):
-        os.environ.update({SCRIPT_FILE_NAME: 'test.py'})
-        self.assertIsNotNone(os.environ.get(SCRIPT_FILE_NAME))
-        self.assertEqual(os.environ.get(SCRIPT_FILE_NAME),
-                         'test.py')
